+++ test/test_basic.ml
@@ -48,7 +48,7 @@ let check_raises_containing name substr f =
     end
 
 let () =
-  check "version is 0.1.105" Version.v "0.1.105";
+  check "version is 0.1.106" Version.v "0.1.106";
 
   (* --- regression --- *)
   check "'1 + 2'"  (Pipeline.process "1 + 2") "3";